# swagger/api-docs-fixed.yaml
swagger: '2.0'
info:
  description: MiriZoom 백엔드 REST API 명세. 인증 API에서 발급받은 Access Token은 Authorize에 'Bearer {token}' 형식으로 입력합니다.
  version: '1.1'
  title: MiriZoom Backend API
host: localhost:8080
basePath: /
...
paths:
...
  /api/tickets:
    post:
      tags:
      - 영업점 대기표 API
      summary: 대기표 발급
      description: 로그인 사용자에게 선택한 영업점과 업무 유형의 대기표를 발급합니다.
      operationId: issueUsingPOST
      consumes:
      - application/json
      produces:
      - '*/*'
      parameters:
      - in: body
        name: request
        description: request
        required: true
        schema:
          $ref: '#/definitions/TicketIssueRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ApiResponseOfTicketIssueResponse'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Bearer: []
      deprecated: false
  /api/tickets/call:
    post:
      tags:
      - 영업점 대기표 API
      summary: 다음 대기표 호출
      description: 선택한 영업점에서 대기 중인 다음 번호를 호출합니다.
      operationId: callNextUsingPOST
      consumes:
      - application/json
      produces:
      - '*/*'
      parameters:
      - in: body
        name: request
        description: request
        required: true
        schema:
          $ref: '#/definitions/TicketCallRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ApiResponseOfTicketCallResponse'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
      - Bearer: []
      deprecated: false
  /api/tickets/status:
    get:
      tags:
      - 영업점 대기표 API
      summary: 영업점 대기 현황 조회
      description: 선택한 영업점의 현재 호출 번호와 대기 인원을 조회합니다.
      operationId: statusUsingGET
      produces:
      - '*/*'
      parameters:
      - name: branchId
        in: query
        description: 영업점 ID
        required: true
        type: integer
        format: int64
        allowEmptyValue: false
        x-example: 1
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ApiResponseOfTicketStatusResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  ApiResponseOfTicketCallResponse:
    type: object
    properties:
      data:
        description: 성공 응답 데이터
        $ref: '#/definitions/TicketCallResponse'
      error:
        type: string
        example: 인증 정보가 없거나 유효하지 않습니다
        description: 실패 메시지
      message:
        type: string
        example: 요청이 정상적으로 처리되었습니다
        description: 성공 메시지
      path:
        type: string
        example: /api/users/me
        description: 요청 URI
      statusCode:
        type: integer
        format: int32
        example: 200
        description: 애플리케이션 응답 코드
      timestamp:
        type: string
        format: date-time
        description: 응답 생성 시각
    title: ApiResponse«TicketCallResponse»
    description: 모든 REST API가 사용하는 공통 응답
  ApiResponseOfTicketIssueResponse:
    type: object
    properties:
      data:
        description: 성공 응답 데이터
        $ref: '#/definitions/TicketIssueResponse'
      error:
        type: string
        example: 인증 정보가 없거나 유효하지 않습니다
        description: 실패 메시지
      message:
        type: string
        example: 요청이 정상적으로 처리되었습니다
        description: 성공 메시지
      path:
        type: string
        example: /api/users/me
        description: 요청 URI
      statusCode:
        type: integer
        format: int32
        example: 200
        description: 애플리케이션 응답 코드
      timestamp:
        type: string
        format: date-time
        description: 응답 생성 시각
    title: ApiResponse«TicketIssueResponse»
    description: 모든 REST API가 사용하는 공통 응답
  ApiResponseOfTicketStatusResponse:
    type: object
    properties:
      data:
        description: 성공 응답 데이터
        $ref: '#/definitions/TicketStatusResponse'
      error:
        type: string
        example: 인증 정보가 없거나 유효하지 않습니다
        description: 실패 메시지
      message:
        type: string
        example: 요청이 정상적으로 처리되었습니다
        description: 성공 메시지
      path:
        type: string
        example: /api/users/me
        description: 요청 URI
      statusCode:
        type: integer
        format: int32
        example: 200
        description: 애플리케이션 응답 코드
      timestamp:
        type: string
        format: date-time
        description: 응답 생성 시각
    title: ApiResponse«TicketStatusResponse»
    description: 모든 REST API가 사용하는 공통 응답
...
  TicketCallRequest:
    type: object
    properties:
      branchId:
        type: integer
        format: int64
    title: TicketCallRequest
  TicketCallResponse:
    type: object
    properties:
      ticketId:
        type: integer
        format: int64
      ticketNumber:
        type: string
    title: TicketCallResponse
  TicketIssueRequest:
    type: object
    properties:
      branchId:
        type: integer
        format: int64
      serviceType:
        type: string
    title: TicketIssueRequest
  TicketIssueResponse:
    type: object
    properties:
      businessDate:
        type: string
        format: date
      ticketId:
        type: integer
        format: int64
      ticketNumber:
        type: string
      waitingCount:
        type: integer
        format: int32
    title: TicketIssueResponse
  TicketStatusResponse:
    type: object
    properties:
      branchId:
        type: integer
        format: int64
      currentCalledNumber:
        type: string
      waitingCount:
        type: integer
        format: int32
    title: TicketStatusResponse
...